% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqRes.r
\name{residqq}
\alias{residqq}
\title{Qq plot of randomized quantile residuals against standard normal quantiles}
\usage{
residqq(umFit, type = "site-sum", main = "Residual qq plot",
  plotLine = TRUE, ...)
}
\arguments{
\item{umFit}{An object from a model fitted using \link[unmarked]{pcount}.}

\item{type}{The type of randomized quantile residual to plot. One of 'site-sum' or 'observation'.}

\item{main}{Plot label.}

\item{plotLine}{If true, the identity line is added to the plot.}

\item{...}{Further arguments passed to qqnorm.}
}
\description{
Qq plot of randomized quantile residuals against standard normal quantiles
}
