% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmixgof.R
\docType{package}
\name{nmixgof}
\alias{nmixgof}
\alias{nmixgof-package}
\title{Goodness of fit checks for binomial N-mixture models}
\description{
The package contains methods to compute overdispersion metrics, randomized quantile residuals,
and graphical diagnostics of model fit for binomial N-mixture models fitted using the \link[unmarked]{unmarked} package.
Details about the checks are given in Knape et al. (2018) and at \url{https://www.biorxiv.org/content/early/2017/09/27/194340}.
}
\references{
Knape et al. 2018. Sensitivity of binomial N-mixture models to overdispersion: 
the importance of assessing model fit. Methods in Ecology and Evolution, in press.
}
