% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmixgof.R
\docType{package}
\name{nmixgof}
\alias{nmixgof}
\alias{nmixgof-package}
\title{Goodness of fit checks for binomial N-mixture models}
\description{
The package contains methods to compute overdispersion metrics, randomized quantile residuals,
and graphical diagnostics of model fit.
}
