% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqRes.r
\name{residcov}
\alias{residcov}
\title{Plot residuals against covariates}
\usage{
residcov(umFit, ...)
}
\arguments{
\item{...}{}
}
\description{
Site-sum randomized quantile residuals are used for site covariates while marginal residuals are used for observation covariates.
The same random residual draws are reused for different covariates.
}
