% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqRes.r
\name{residfit}
\alias{residfit}
\title{Plot residuals against fitted values}
\usage{
residfit(umFit, type = "marginal", residual = "rq", ...)
}
\arguments{
\item{umFit}{An object from a model fitted using \link[unmarked]{pcount}.}

\item{type}{The type of randomized quantile residual to plot. One of 'margina', 'site-sum' or 'observation'.}

\item{residual}{The type of residual. Only 'rq' currently implemented.}

\item{...}{Plot arguments.}
}
\description{
Plots randomized-quantile residuals for binomial N-mixture models against fitted values.
}
